+++ lib/txbr/liquid.rb
@@ -1,0 +1,5 @@
+module Txbr
+  module Liquid
+    autoload :ConnectedContentTag, 'txbr/liquid/connected_content_tag'
+  end
+end