+++ spec/email_templates_api_spec.rb
@@ -1,55 +1,18 @@
 require 'spec_helper'
 require 'support/fake_connection'
+require 'shared_examples/api_errors'
 
-describe Txbr::BrazeApi do
+require 'json'
+
+describe Txbr::EmailTemplatesApi do
   let(:api_key) { 'abc123' }
   let(:api_url) { 'https://somewhere.braze.com' }
   let(:connection) { FakeConnection.new(interactions) }
-  let(:client) { described_class.new(api_key, api_url, connection: connection) }
+  let(:braze_api) { Txbr::BrazeApi.new(api_key, api_url, connection: connection) }
+  let(:client) { described_class.new(braze_api) }
 
-  shared_examples 'a client request that handles errors' do
-    context 'when the resource is not found' do
-      let(:interactions) do
-        super().tap do |inter|
-          inter[0][:response][:status] = 404
-        end
-      end
-
-      it 'raises a not found error' do
-        expect { subject }.to raise_error(Txbr::BrazeNotFoundError)
-      end
-    end
-
-    context 'when the request is unauthorized' do
-      let(:interactions) do
-        super().tap do |inter|
-          inter.unshift(
-            request: inter[0][:request],
-            response: { status: 401 }
-          )
-        end
-      end
-
-      it 'raises an unauthorized error' do
-        expect { subject }.to raise_error(Txbr::BrazeUnauthorizedError)
-      end
-    end
-
-    context 'when some other bad thing happens' do
-      let(:interactions) do
-        super().tap do |inter|
-          inter[0][:response][:status] = 500
-        end
-      end
-
-      it 'raises a generic error' do
-        expect { subject }.to raise_error(Txbr::BrazeApiError)
-      end
-    end
-  end
-
-  describe '#each_email_template' do
-    subject { client.each_email_template.to_a }
+  describe '#each' do
+    subject { client.each.to_a }
 
     before do
       stub_const("#{described_class.name}::TEMPLATE_BATCH_SIZE", 1)
@@ -78,8 +41,8 @@ describe Txbr::BrazeApi do
     it_behaves_like 'a client request that handles errors'
   end
 
-  describe '#get_email_template_details' do
-    subject { client.get_email_template_details(email_template_id: email_template_id) }
+  describe '#details' do
+    subject { client.details(email_template_id: email_template_id) }
     let(:email_template_id) { 'abc123' }
 
     let(:details) do
@@ -93,7 +56,7 @@ describe Txbr::BrazeApi do
 
     let(:interactions) do
       [{
-        request: { verb: 'get', url: described_class::TEMPLATE_INFO_PATH, params: { email_template_id: email_template_id } },
+        request: { verb: 'get', url: described_class::TEMPLATE_DETAILS_PATH, params: { email_template_id: email_template_id } },
         response: { status: 200, body: details.to_json }
       }]
     end